agentcore/requirements.txt
bedrock-agentcore
strands-agents
boto3
mcp
